backend/requirements.txt
fastapi>=0.129.0
uvicorn[standard]>=0.41.0
sqlalchemy>=2.0.46
pydantic[email]>=2.12.5
pydantic-settings>=2.13.0
asyncpg>=0.31.0
aiosqlite>=0.19.0
redis[hiredis]>=5.0.0
websockets>=16.0
python-dotenv>=1.2.1
agentscope>=1.0.18
openai>=2.21.0
python-multipart>=0.0.22
aiofiles>=23.1.0
loguru>=0.7.3
greenlet>=3.3.1
alembic>=1.18.4
psycopg2-binary>=2.9.9
requests>=2.31.0
Pillow>=12.1.1
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
google-genai>=1.65.0
xai-sdk>=1.7.0
volcengine-python-sdk[ark]>=1.0.0
httpx>=0.27.0
python-frontmatter>=1.0.0
packaging>=23.0
user-agents>=2.2.0
cachetools>=7.0.6
cryptography>=47.0.0
# --- Phase1+ infra ---
aioboto3>=12.3.0
arq>=0.26.0
purgatory>=3.0.0
slowapi>=0.1.9
